# openapi/openapi.json
{
  "components": {
    "schemas": {
      "Error": {
        "additionalProperties": false,
        "description": "The single error shape returned by every Acme Commerce API error response.",
        "example": {
          "error": {
            "code": "VALIDATION_ERROR",
            "details": {
              "fields": [
                {
                  "field": "body.title",
                  "message": "\"title\" is required.",
                  "rule": "required"
                }
              ]
            },
            "message": "The request body is invalid.",
            "request_id": "req_18f3a9c4e21b7d05f6a3b8c1"
          }
        },
        "properties": {
          "error": {
            "additionalProperties": false,
            "properties": {
              "code": {
                "description": "Stable machine-readable error code. Branch on this, never on `message`. The full list is documented in README.md.",
                "example": "SKU_ALREADY_EXISTS",
                "type": "string"
              },
              "details": {
                "additionalProperties": true,
                "description": "Optional structured context. Its shape depends on `code`: `VALIDATION_ERROR` carries a `fields` array, `MALFORMED_ID` carries `expected_pattern`, `SKU_ALREADY_EXISTS` carries the conflicting `sku`.",
                "properties": {},
                "type": "object"
              },
              "message": {
                "description": "Human-readable explanation, for a log or a response pane. May be reworded in any release; it is not part of the contract.",
                "type": "string"
              },
              "request_id": {
                "description": "Correlation identifier for this request, also returned in the `X-Request-Id` response header. Quote it when reporting a problem — it is what locates the server-side log lines.",
                "example": "req_18f3a9c4e21b7d05f6a3b8c1",
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "request_id"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "title": "Error",
        "type": "object"
      },
...
      "ProductCreate": {
        "additionalProperties": false,
        "description": "Body for creating a product. Unknown properties are REJECTED with 400 rather than ignored — a silently-dropped `titel` is a typo that ships to production believing it worked.",
        "example": {
          "description": "A 2.5-layer shell that stuffs into its own pocket.",
          "product_type": "Apparel",
          "status": "active",
          "tags": [
            "waterproof",
            "lightweight"
          ],
          "title": "Riverbend Packable Rain Jacket",
          "vendor": "Acme"
        },
        "properties": {
          "description": {
            "description": "Optional long-form description.",
            "maxLength": 5000,
            "type": "string"
          },
          "product_type": {
            "description": "Optional merchandising category.",
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "status": {
            "default": "draft",
            "description": "Optional. Defaults to `draft`, so a new product is not accidentally live.",
            "enum": [
              "draft",
              "active",
              "archived"
            ],
            "type": "string"
          },
          "tags": {
            "description": "Optional labels. Defaults to an empty array.",
            "items": {
              "maxLength": 50,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 50,
            "type": "array"
          },
          "title": {
            "description": "Required. Customer-facing product name.",
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "vendor": {
            "description": "Optional brand or supplier.",
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "ProductCreate",
        "type": "object"
      },
      "ProductListResponse": {
        "additionalProperties": false,
        "description": "A page of products plus the pagination metadata needed to fetch the rest.",
        "properties": {
          "data": {
            "description": "The requested page of products. Empty when nothing matches.",
            "items": {
              "$ref": "#/components/schemas/Product"
            },
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "required": [
          "data",
          "pagination"
        ],
        "title": "ProductListResponse",
        "type": "object"
      },
      "ProductResponse": {
        "additionalProperties": false,
        "description": "A single product.",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Product"
          }
        },
        "required": [
          "data"
        ],
        "title": "ProductResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "A development bearer token from `POST /api/v1/auth/token`.\n\nSend it as `Authorization: Bearer <token>`.\n\nThe token is a JWT: signed, **not encrypted**. Anyone holding it can read every claim inside it, so treat it as a credential and keep it out of anything you commit. It cannot be revoked individually — it is valid until it expires.\n\nCatalog **reads** are deliberately unauthenticated: a storefront browses the catalog without credentials. Everything that writes, and everything under locations, inventory, and pricing, requires a token.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
...
  "info": {
    "contact": {
      "email": "platform@acme.example",
      "name": "Acme Commerce Platform Team"
    },
...
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "summary": "Internal ecommerce platform API for catalog, inventory, orders, and fulfillment.",
    "title": "Acme Commerce API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "description": "Returns a page of products with pagination metadata.\n\n**No implicit filtering.** With no `status` parameter, products of every status are\nreturned, archived ones included. This API applies no hidden default filter: the result\nset is fully determined by the query string. Most catalog APIs hide archived records;\nthis one does not, so that DELETE is observable as a state change rather than as a\ndisappearance.\n\nFilters combine with AND. An empty result is `200` with `\"data\": []`, never `404` —\nan empty collection is a successful answer to a well-formed question. A page past the\nend is also `200` with an empty array.\n\n**Unknown query parameters are rejected with 400**, not ignored. `?statuss=active`\nsilently returning every product is a worse outcome than a loud error.\n\nOrdering always ends with `id` as a tiebreaker, so paging through a sorted list cannot\nshow the same record twice.\n\nEvery response carries an `X-Request-Id` header.",
        "operationId": "listProducts",
        "parameters": [
          {
            "description": "One-based page number. Requesting a page past the end returns an empty `data` array with 200, not 404.",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Records per page. Default 25, maximum 100. A larger value is rejected with 400 rather than silently clamped, so a client asking for 1000 learns that it cannot have 1000.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Filter by publication state. NOTE: with no `status` filter, products of ALL statuses are returned, archived included. This API applies no hidden default filter — the result set is fully determined by the query string.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "draft",
                "active",
                "archived"
              ],
              "type": "string"
            }
          },
          {
            "description": "Filter by vendor. Case-insensitive EXACT match, not a prefix or substring.",
            "example": "Acme",
            "in": "query",
            "name": "vendor",
            "required": false,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "description": "Filter by product type. Case-insensitive exact match.",
            "example": "Backpacks",
            "in": "query",
            "name": "product_type",
            "required": false,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "description": "Filter to products carrying this tag. Case-sensitive exact match on an array element. One tag per request in this version.",
            "example": "bestseller",
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "maxLength": 50,
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "description": "Case-insensitive substring search across `title`, `description`, and `vendor`. Does not search variant SKUs. Combines with the filters using AND.",
            "example": "backpack",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "description": "Field to sort by. Restricted to this allowlist; any other value returns 400 with the permitted values in `error.details.allowed`. `id` is always appended as a tiebreaker so paging is stable across requests.",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "created_at",
              "enum": [
                "created_at",
                "updated_at",
                "title",
                "vendor",
                "product_type",
                "status"
              ],
              "type": "string"
            }
          },
          {
            "description": "Sort direction. Applies to the tiebreaker too.",
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "description": "Optional caller-supplied correlation id, echoed back in the X-Request-Id response header and included in structured errors. When absent or not matching this pattern, the server generates one. Supplying your own lets you correlate a request across your system and ours.",
            "example": "my-app-checkout-attempt-42",
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "schema": {
              "pattern": "^[A-Za-z0-9_-]{1,128}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponse"
                }
              }
            },
            "description": "A page of products."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is malformed. `error.code` is VALIDATION_ERROR, MALFORMED_ID, or INVALID_JSON."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "An unexpected server error. `error.request_id` identifies the failure in the server logs."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The database is unreachable. Safe to retry; see the Retry-After header."
          }
        },
        "summary": "List products",
        "tags": [
          "Catalog"
        ]
      },
      "post": {
        "description": "Creates a product. The identifier, timestamps, and archive state are assigned by the\nserver and cannot be supplied.\n\nA new product defaults to `draft` status so that creating one does not accidentally\npublish it.\n\nUnknown body properties are rejected with `400` rather than ignored, so a misspelled\nfield is visible immediately instead of shipping to production believing it worked.\n\nResponds `201` with a `Location` header pointing at the new product.",
        "operationId": "createProduct",
        "parameters": [
          {
            "description": "Optional caller-supplied correlation id, echoed back in the X-Request-Id response header and included in structured errors. When absent or not matching this pattern, the server generates one. Supplying your own lets you correlate a request across your system and ours.",
            "example": "my-app-checkout-attempt-42",
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "schema": {
              "pattern": "^[A-Za-z0-9_-]{1,128}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            },
            "description": "The product that was created."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is malformed. `error.code` is VALIDATION_ERROR, MALFORMED_ID, or INVALID_JSON."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Missing, malformed, invalid, or expired bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Authenticated, but your role lacks `catalog:write` (`INSUFFICIENT_PERMISSION`)."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "An unexpected server error. `error.request_id` identifies the failure in the server logs."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The database is unreachable. Safe to retry; see the Retry-After header."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a product",
        "tags": [
          "Catalog"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "The server this document was fetched from.",
      "url": "/"
    }
  ],
...
}